feat: Add LottoController to separate application flow control
